Refactor: AuthFormLayout 생성 - 회원가입, 로그인 폼에 공통적으로 활용
