+++ src/pages/Login.tsx
@@ -1,10 +1,15 @@
 import { useNavigate } from 'react-router';
-import { Link } from 'react-router-dom';
 import RectButton from '../components/common/RectButton';
+import AuthFormLayout from '../components/layout/AuthFormLayout';
 import useInput from '../hooks/useInput';
 import { UserType } from '../types';
 
 type LoginProps = {};
+
+const linkOption = {
+  to: '/join',
+  text: '회원가입 하러가기',
+};
 
 export default function Login({}: LoginProps) {
   const navigate = useNavigate();
@@ -53,33 +58,25 @@ export default function Login({}: LoginProps) {
   };
 
   return (
-    <div className="@w(~375):w(150) @w(376~):w(360) h(240)  m(auto/auto) p(15) bg(white) r(10) vbox(center)">
-      <h1 className="font(24) bold">미니타닥 로그인</h1>
-      <div className="space(60)"></div>
-      <form onSubmit={onSubmitForm} className="hbox gap(10)">
-        <div className="vbox(right) gap(10)">
-          <div>
-            <label htmlFor="user_id">아이디</label>
-            <input value={userId} onChange={onChangeUserId} id="user_id" className="m(0/5/0/5)" maxLength={12} />
-          </div>
-          <div>
-            <label htmlFor="user_password">비밀번호</label>
-            <input
-              type="password"
-              value={password}
-              onChange={onChangePassword}
-              id="user_password"
-              className="m(0/5/0/5)"
-              maxLength={12}
-            />
-          </div>
+    <AuthFormLayout title={'미니타닥 로그인'} link={linkOption} onSubmitForm={onSubmitForm}>
+      <div className="vbox(right) gap(10)">
+        <div>
+          <label htmlFor="user_id">아이디</label>
+          <input value={userId} onChange={onChangeUserId} id="user_id" className="m(0/5/0/5)" maxLength={12} />
         </div>
-        <RectButton buttonName="로그인" w="70" h="70" />
-      </form>
-      <div className="space(10)"></div>
-      <Link to="/join" className="c(blue)">
-        회원가입 하러가기
-      </Link>
-    </div>
+        <div>
+          <label htmlFor="user_password">비밀번호</label>
+          <input
+            type="current-password"
+            value={password}
+            onChange={onChangePassword}
+            id="user_password"
+            className="m(0/5/0/5)"
+            maxLength={12}
+          />
+        </div>
+      </div>
+      <RectButton buttonName="로그인" w="70" h="70" />
+    </AuthFormLayout>
   );
 }
